Add support for `rails@5.1`, drop `rails@4.1`
With the release of `rails@5.1`, add explicit test coverage, in addition
to testing on `master`.

According to the [Ruby on Rails Maintenance Policy][policy], Rails 4.1
has reached its end of life stage [after the release of Rails
5.0][rails-5.0].

As such, this commit drops formal test coverage and support for Rails
versions prior to 4.2.

[policy]: http://guides.rubyonrails.org/maintenance_policy.html#severe-security-issues
[rails-5.0]: http://weblog.rubyonrails.org/2016/6/30/Rails-5-0-final/
[rails-5.1]: http://weblog.rubyonrails.org/2017/4/27/Rails-5-1-final/