Redirect with a trailing slash when needed
For apps not mounted at `/`, apps need to write custom redirect logic to
add the trailing slash that Ember requires.

For example, if an app is mounted at `/admin`, any requests to `/admin`
will be redirected to `/admin/` because [Ember requires a trailing slash
when visiting the root URL][issue]. Otherwise, the
request isn't affected.

This commit implements that logic as part of `mount_ember_app`.

[issue]: https://github.com/emberjs/ember.js/issues/4751